+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-12-21
+### Added
+- Minimum and maximum of each minor versions of Unity.
+
+### Changed
+- Menu item is now in Window instead of Window/NG Tools.
+- Order of results is now sorted by version.
+
+### Removed
+- Removed unused menu items.
+
 ## [0.2.0] - 2019-12-19
 ### Added
 - Export all in the context menu.